Create Valuable time
Valuable Time
Ravi is asked to solve a challenge by his friend Salma. He is given a sequence and is asked to perform operations on it such that the sequence becomes the same as Salma's sequence in minimal time, and he need to clear the sequence during the process.

In rotation operation, Ravi can remove the first element of his sequence and push it in the end, it takes 
1
 unit of time.
In clear operation, an element can be removed from the front, each remove costs 
1
 unit of time for Ravi's list. It is not necessary that clearing operation will be performed only after the sequence becomes the same, we can also clear if the front element of both sequences are the same. The clear operation will be performed on both the sequences and for Salma's sequence it will take zero amount of time.
Suppose the sequence of numbers given to Ravi and Salma are 
3
−
2
−
1
 and 
1
−
3
−
2
 respectively, then the ways in which Ravi's sequence can be rearranged to form Salma's sequence one can be explained as below:

In the first iteration since 
3
 and 
1
 are unequal so, we rotate to make the first sequence 
2
−
1
−
3
 with a time of 
1
 unit.
In the second iteration the first integer 
2
 and 
1
 of both the sequences respectively are unequal so, again rotate the integer of the first sequence to make it 
1
−
3
−
2
 with a time of 
1
 unit.
In the third iteration, the first element of both sequences is equal so we can perform the clear operation. It will take 
1
 unit of time. Now Ravi's sequence is 
3
−
2
 and Salma's sequence is 
3
−
2
.
In the fourth iteration, the list becomes 
2
 with a time of 
1
 unit.
And, finally in the fifth iteration the last element 
2
 is cleared with a time of 
1
 unit.
You have to find the total time required to complete the process.

Input Format
The first line contains an integer 
N
 which denotes the total number of integers given to Ravi and Salma.
The second line contains 
N
 integers given to Ravi as the sequence.
The third line contains the 
N
 integers given to Salma as the sequence.     

Output Format
Print the total time required to complete the process.

Constraints
1
<=
N
<=
50

0
<=
R
<=
100

0
<=
S
<=
100

Time Limit
1 second

Example
Input
5
5 4 2 3 1
5 2 1 4 3

Output
7

Sample test case explanation
We can clear the first element as it is the same. Remaining sequence of Ravi is 
[
4
,
2
,
3
,
1
]
 and Salma is 
[
2
,
1
,
4
,
3
]
.
Since the first of both lists are different, perform the rotation operation and Ravi's sequence will become 
[
2
,
3
,
1
,
4
]
.
We can clear the first element as it is the same. The remaining sequence of Ravi is 
[
3
,
1
,
4
]
, and Salma is 
[
1
,
4
,
3
]
.
Since the first of both lists are different, perform the rotation operation and Ravi's sequence will become 
[
1
,
4
,
3
]
.
We can clear the first element as it is the same. The remaining sequence of Ravi is 
[
4
,
3
]
 and Salma is 
[
4
,
3
]
.
We can clear the first element as it is the same. The remaining sequence of Ravi is 
[
3
]
 and Salma is 
[
3
]
.
We can clear the first element as it is the same. The remaining sequence of Ravi is 
[
]
 and Salma is 
[
]
.
So a total of 
7
 units of time is taken.